adjustments for Windows compilation

+++ include/JasonBuilder.h
@@ -302,11 +302,11 @@ namespace arangodb {
         void addInt (int64_t v) {
           if (v >= 0 && v <= 9) {
             reserveSpace(1);
-            _start[_pos++] = 0x30 + v;
+            _start[_pos++] = static_cast<uint8_t>(0x30 + v);
           }
           else if (v < 0 && v >= -6) {
             reserveSpace(1);
-            _start[_pos++] = 0x40 + v;
+            _start[_pos++] = static_cast<uint8_t>(0x40 + v);
           }
           else {
             appendInt(v, 0x1f);
@@ -316,7 +316,7 @@ namespace arangodb {
         void addUInt (uint64_t v) {
           if (v <= 9) {
             reserveSpace(1);
-            _start[_pos++] = 0x30 + v;
+            _start[_pos++] = static_cast<uint8_t>(0x30 + v);
           }
           else {
             appendUInt(v, 0x27);
@@ -341,7 +341,7 @@ namespace arangodb {
           }
           else {
             // short string
-            _start[_pos++] = 0x40 + strLen;
+            _start[_pos++] = static_cast<uint8_t>(0x40 + strLen);
           }
           target = _start + _pos;
           _pos += strLen;
